Extract external link into component
